commit changes tables
