+++ src/Entity/Cargo.php
@@ -2,7 +2,6 @@
 
 namespace App\Entity;
 
-
 use Doctrine\ORM\Mapping as ORM;
 
 /**
@@ -28,6 +27,26 @@ class Cargo
      * @ORM\Column(name="descripcion", type="string", length=250, nullable=false)
      */
     private $descripcion;
+    
+    // Getters
 
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function getDescripcion(): ?string
+    {
+        return $this->descripcion;
+    }
+
+    // Setters
+
+    public function setDescripcion(string $descripcion): self
+    {
+        $this->descripcion = $descripcion;
+
+        return $this;
+    }
 
 }
