--- src/Entity/Cliente.php
@@ -1,105 +1,0 @@
-<?php
-
-namespace App\Entity;
-
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ORM\Entity
- */
-class Cliente
-{
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
-    private $id;
-
-    /**
-     * @ORM\Column(type="string", length=255, unique=true)
-     */
-    private $nombre;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    private $apellidos;
-
-    /**
-     * @ORM\Column(type="string", length=255, unique=true)
-     */
-    private $email;
-
-    /**
-     * @ORM\Column(type="datetime")
-     */
-    private $fechaNacimiento;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    private $direccion;
-
-    // Añade los getters y setters para cada propiedad
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getNombre(): ?string
-    {
-        return $this->nombre;
-    }
-
-    public function setNombre(string $nombre): self
-    {
-        $this->nombre = $nombre;
-        return $this;
-    }
-
-    public function getApellidos(): ?string
-    {
-        return $this->apellidos;
-    }
-
-    public function setApellidos(string $apellidos): self
-    {
-        $this->apellidos = $apellidos;
-        return $this;
-    }
-
-    public function getEmail(): ?string
-    {
-        return $this->email;
-    }
-
-    public function setEmail(string $email): self
-    {
-        $this->email = $email;
-        return $this;
-    }
-
-    public function getFechaNacimiento(): ?\DateTimeInterface
-    {
-        return $this->fechaNacimiento;
-    }
-
-    public function setFechaNacimiento(\DateTimeInterface $fechaNacimiento): self
-    {
-        $this->fechaNacimiento = $fechaNacimiento;
-        return $this;
-    }
-
-    public function getDireccion(): ?string
-    {
-        return $this->direccion;
-    }
-
-    public function setDireccion(string $direccion): self
-    {
-        $this->direccion = $direccion;
-        return $this;
-    }
-}
