--- src/Entity/Departamento.php
@@ -1,41 +1,0 @@
-<?php
-
-namespace App\Entity;
-
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ORM\Entity
- */
-class Departamento
-{
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
-    private $id_Departamento;
-
-    /**
-     * @ORM\Column(type="string", length=255, unique=true)
-     */
-    private $nombre;
-
-
-
-    public function getId(): ?int
-    {
-        return $this->id_Departamento;
-    }
-
-    public function getNombre(): ?string
-    {
-        return $this->nombre;
-    }
-
-    public function setNombre(string $nombre): self
-    {
-        $this->nombre = $nombre;
-        return $this;
-    }
-}
