--- src/Entity/Usuario3.php
@@ -1,153 +1,0 @@
-<?php
-
-namespace App\Entity;
-
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping as ORM;
-use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
-
-/**
- * @ORM\Entity(repositoryClass="App\Repository\UsuarioRepository")
- * @UniqueEntity(fields={"email"}, message="El correo electrónico ya está en uso")
- */
-class Usuario
-{
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
-    private $id;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    private $nombre;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    private $apellidos;
-
-    /**
-     * @ORM\Column(type="string", length=255, unique=true)
-     */
-    private $email;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    private $password;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    private $rol;
-
-    /**
-     * @ORM\OneToMany(targetEntity="App\Entity\Incidencia", mappedBy="usuario", cascade={"persist", "remove"})
-     */
-    private $incidencias;
-
-    public function __construct()
-    {
-        $this->incidencias = new ArrayCollection();
-    }
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getNombre(): ?string
-    {
-        return $this->nombre;
-    }
-
-    public function setNombre(string $nombre): self
-    {
-        $this->nombre = $nombre;
-
-        return $this;
-    }
-
-    public function getApellidos(): ?string
-    {
-        return $this->apellidos;
-    }
-
-    public function setApellidos(string $apellidos): self
-    {
-        $this->apellidos = $apellidos;
-
-        return $this;
-    }
-
-    public function getEmail(): ?string
-    {
-        return $this->email;
-    }
-
-    public function setEmail(string $email): self
-    {
-        $this->email = $email;
-
-        return $this;
-    }
-
-    public function getPassword(): ?string
-    {
-        return $this->password;
-    }
-
-    public function setPassword(string $password): self
-    {
-        $this->password = $password;
-
-        return $this;
-    }
-
-    public function getRol(): ?string
-    {
-        return $this->rol;
-    }
-
-    public function setRol(string $rol): self
-    {
-        $this->rol = $rol;
-
-        return $this;
-    }
-
-    /**
-     * @return Collection|Incidencia[]
-     */
-    public function getIncidencias(): Collection
-    {
-        return $this->incidencias;
-    }
-
-    public function addIncidencia(Incidencia $incidencia): self
-    {
-        if (!$this->incidencias->contains($incidencia)) {
-            $this->incidencias[] = $incidencia;
-            $incidencia->setUsuario($this);
-        }
-
-        return $this;
-    }
-
-    public function removeIncidencia(Incidencia $incidencia): self
-    {
-        if ($this->incidencias->contains($incidencia)) {
-            $this->incidencias->removeElement($incidencia);
-            // set the owning side to null (unless already changed)
-            if ($incidencia->getUsuario() === $this) {
-                $incidencia->setUsuario(null);
-            }
-        }
-
-        return $this;
-    }
-}
